ordercheck 변경0714

+++ orderCheck.py
@@ -5,8 +5,9 @@ import models
 import datetime
 from pybithumb import Bithumb
 
-secretKey = "07c1879d34d18036405f1c4ae20d3023"
-connenctKey = "9ae8ae53e7e0939722284added991d55"
+secretKey = "ee7741a2e52957613c020ded3c91751c"
+connenctKey = "ef3d9e8fb9b15ca740150fed18cdaaae"
+
 bithumb = Bithumb(connenctKey, secretKey)
 
 now1 = datetime.datetime.now()
@@ -16,7 +17,8 @@ try:
 finally:
     db.close()
 
-autoStatus = db.query(models.autoTradingStatus).filter(models.autoTradingStatus.status == 1).first()
+autoStatus = db.query(models.autoTradingStatus).filter(
+    models.autoTradingStatus.status == 1).first()
 
 if autoStatus == None:
     print('자동 매매 정지 exit')
@@ -31,26 +33,41 @@ for order in orderList:
     if order.status == 1:
         isOrder = []
 
-        order_desc = ['bid',order.coin, order.order_id, 'KRW']
+        order_desc = ['bid', order.coin, order.order_id, 'KRW']
         orderStatus = bithumb.get_order_completed(order_desc)
         if orderStatus['data']['order_status'] == 'Completed':
-            had_coin = db.query(models.possessionCoin).filter(models.possessionCoin.coin == order.coin).first()
+            had_coin = db.query(models.possessionCoin).filter(
+                models.possessionCoin.coin == order.coin).first()
 
-            had_coin.unit = float(had_coin.unit) + float(orderStatus['data']['contract'][0]['units'])
-            had_coin.price = float(had_coin.price) + float(orderStatus['data']['contract'][0]['price'])
-            had_coin.total = float(had_coin.total) + float(orderStatus['data']['contract'][0]['total'])
-            had_coin.fee = float(had_coin.fee) + float(orderStatus['data']['contract'][0]['fee'])
+            orderSum = {'unit': 0, 'total': 0, 'fee': 0}
+
+            for cont in orderStatus['data']['contract']:
+                orderSum['unit'] += float(cont['units'])
+                orderSum['total'] += float(cont['total'])
+                orderSum['fee'] += float(cont['fee'])
+
+            had_coin.unit = float(
+                had_coin.unit) + orderSum['unit']
+            had_coin.price = float(
+                had_coin.price) + float(orderStatus['data']['order_price'])
+            had_coin.total = float(
+                had_coin.total) + orderSum['total']
+            had_coin.fee = float(had_coin.fee) + \
+                orderSum['fee']
+
             had_coin.status = 0
             had_coin.transaction_time = datetime.datetime.now()
+            had_coin.trailingstop_flag = 0
+            had_coin.max = had_coin.price
 
             db.delete(order)
 
             transactionLog = models.possessionLog()
             transactionLog.coin = order.coin
-            transactionLog.unit = orderStatus['data']['contract'][0]['units']
+            transactionLog.unit = orderSum['unit']
             transactionLog.price = orderStatus['data']['contract'][0]['price']
-            transactionLog.total = orderStatus['data']['contract'][0]['total']
-            transactionLog.fee = orderStatus['data']['contract'][0]['fee']
+            transactionLog.total = orderSum['total']
+            transactionLog.fee = orderSum['fee']
             transactionLog.status = 0
             transactionLog.transaction_time = order.transaction_time
             transactionLog.conclusion_time = datetime.datetime.now()
@@ -59,15 +76,40 @@ for order in orderList:
             db.add(transactionLog)
 
         if orderStatus['data']['order_status'] == 'Pending':
-            print(datetime.datetime.utcfromtimestamp(int(orderStatus['data']['order_date'][:-6]) + krTime))
-            timeCheck = str(datetime.datetime.strptime(order.cancel_time, '%Y-%m-%d %H:%M:%S.%f') - datetime.datetime.now())
+            orderSum = {'unit': 0, 'total': 0, 'fee': 0}
+            if len(orderStatus['data']['contract']) > 0:
+                for cont in orderStatus['data']['contract']:
+                    orderSum['unit'] += cont['units']
+                    orderSum['total'] += cont['total']
+                    orderSum['fee'] += cont['fee']
+
+                orderSum['price'] = orderStatus['data']['order_price']
+
+                trading = db.query(models.possessionCoin).filter(
+                    models.possessionCoin.coin == orderStatus['data']['order_currency'])
+
+                trading.unit = orderSum['unit']
+                trading.total = orderSum['total']
+                trading.price = orderSum['price']
+                trading.fee = orderSum['fee']
+
+                db.commit()
+
+            print(datetime.datetime.utcfromtimestamp(
+                int(orderStatus['data']['order_date'][:-6]) + krTime))
+            timeCheck = str(datetime.datetime.strptime(
+                order.cancel_time, '%Y-%m-%d %H:%M:%S.%f') - datetime.datetime.now())
+
             print(timeCheck[0])
             if timeCheck[0] == '-':
-                delpossession = db.query(models.possessionCoin).filter(models.possessionCoin.coin == order.coin).first()
+                delpossession = db.query(models.possessionCoin).filter(
+                    models.possessionCoin.coin == order.coin).first()
+
                 cancel = bithumb.cancel_order(order_desc)
                 if cancel == True:
                     db.delete(order)
-                    db.delete(delpossession)
+                    if len(orderStatus['data']['contract']) != 0:
+                        db.delete(delpossession)
 
         print(orderStatus)
 
@@ -77,27 +119,35 @@ for order in orderList:
     if order.status == 3 or order.status == 5:
         isOrder = []
 
-        order_desc = ['ask',order.coin, order.order_id, 'KRW']
+        order_desc = ['ask', order.coin, order.order_id, 'KRW']
         orderStatus = bithumb.get_order_completed(order_desc)
-        Possession = db.query(models.possessionCoin).filter(models.possessionCoin.coin == order.coin).first()
+        Possession = db.query(models.possessionCoin).filter(
+            models.possessionCoin.coin == order.coin).first()
+
         if orderStatus['data']['order_status'] == 'Completed':
-            poseesionCheck = db.query(models.orderCoin).filter(models.orderCoin.coin == order.coin).all()
-            if len(poseesionCheck) == 1:
-                db.delete(Possession)
-            else:
-                Possession.unit = float(Possession.unit) - float(orderStatus['data']['contract'][0]['units'])
-                Possession.unit = float(Possession.price) - float(orderStatus['data']['contract'][0]['price'])
-                Possession.unit = float(Possession.total) - float(orderStatus['data']['contract'][0]['total'])
-                Possession.unit = float(Possession.fee) - float(orderStatus['data']['contract'][0]['fee'])
-                Possession.transaction_time = order.transaction_time
+            orderSum = {'unit': 0, 'total': 0, 'fee': 0}
+
+            for cont in orderStatus['data']['contract']:
+                orderSum['unit'] += float(cont['units'])
+                orderSum['total'] += float(cont['total'])
+                orderSum['fee'] += float(cont['fee'])
+            orderSum['price'] = orderStatus['data']['order_price']
+
+            poseesionCheck = db.query(models.orderCoin).filter(
+                models.orderCoin.coin == order.coin).all()
+            db.delete(Possession)
             db.delete(order)
+
+            for cont in orderStatus['data']['contract']:
+                print(
+                    'asdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasdasd')
 
             transactionLog = models.possessionLog()
             transactionLog.coin = order.coin
-            transactionLog.unit = orderStatus['data']['contract'][0]['units']
-            transactionLog.price = orderStatus['data']['contract'][0]['price']
-            transactionLog.total = orderStatus['data']['contract'][0]['total']
-            transactionLog.fee = orderStatus['data']['contract'][0]['fee']
+            transactionLog.unit = orderSum['unit']
+            transactionLog.price = orderSum['price']
+            transactionLog.total = orderSum['total']
+            transactionLog.fee = orderSum['fee']
             transactionLog.status = 6
             transactionLog.transaction_time = order.transaction_time
             transactionLog.conclusion_time = datetime.datetime.now()
@@ -107,9 +157,34 @@ for order in orderList:
             db.add(transactionLog)
 
         if orderStatus['data']['order_status'] == 'Pending':
-            timeCheck = str(datetime.datetime.strptime(order.cancel_time, '%Y-%m-%d %H:%M:%S.%f') - datetime.datetime.now())
+            orderSum = {'unit': 0, 'total': 0, 'fee': 0}
+            if len(orderStatus['data']['contract']) > 0:
+                for cont in orderStatus['data']['contract']:
+                    orderSum['unit'] += cont['units']
+                    orderSum['total'] += cont['total']
+                    orderSum['fee'] += cont['fee']
+
+                orderSum['price'] = orderStatus['data']['order_price']
+
+                trading = db.query(models.possessionCoin).filter(
+                    models.possessionCoin.coin == orderStatus['data']['order_currency'])
+
+                trading.unit = trading.unit - orderSum['unit']
+                trading.total = trading.total - orderSum['total']
+                trading.price = orderSum['price']
+                trading.fee = trading.fee - orderSum['fee']
+
+                db.commit()
+
+            timeCheck = str(datetime.datetime.strptime(
+                order.cancel_time, '%Y-%m-%d %H:%M:%S.%f') - datetime.datetime.now())
             print(timeCheck[0])
+
             if timeCheck[0] == '-':
+                if len(orderStatus['data']['contract']) > 0:
+                    if trading.total < 1000:
+                        continue
+
                 cancel = bithumb.cancel_order(order_desc)
                 if cancel == True:
                     db.delete(order)
